requirements.txt
langchain
langchain-community
python-dotenv
langchain-google-genai
streamlit
langchain-groq
langchain_core
sentence_transformers